# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API WeSwing",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost/M14/Projecte_Final/grup2/server/public/index.php"
        }
    ],
    "paths": {
...
        "/api/agrupacions": {
            "get": {
                "tags": [
                    "Agrupacions"
                ],
                "summary": "Veure totes les agrupacions.",
                "operationId": "App\\Http\\Controllers\\AgrupacioController::getAgrupacions",
                "responses": {
                    "200": {
                        "description": "Retorna totes les agrupacions."
                    },
                    "default": {
                        "description": "S'ha produit un error."
                    }
                }
            }
        },
        "/api/agrupacio/{id}": {
            "get": {
                "tags": [
                    "Agrupacions"
                ],
                "summary": "Veure l'agrupacio desitjada.",
                "description": "",
                "operationId": "App\\Http\\Controllers\\AgrupacioController::getAgrupacio",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Agrupacio per id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Agrupacio amb id seleccionat"
                    },
                    "500": {
                        "description": "S'ha produit un error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Agrupacions"
                ],
                "summary": "Actualitza agrupacio ja existent",
                "description": "",
                "operationId": "updateAgrupacio",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id agrupacio a modificar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "nom",
                        "in": "query",
                        "description": "Nom",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "descripcio",
                        "in": "query",
                        "description": "Descripcio",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Agrupacio modificada"
                    },
                    "500": {
                        "description": "S'ha produit un error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Agrupacions"
                ],
                "summary": "Eliminar una agrupacio.",
                "description": "",
                "operationId": "App\\Http\\Controllers\\AgrupacioController::eliminarAgrupacio",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Arupacio a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Agrupacio eliminada"
                    },
                    "500": {
                        "description": "S'ha produit un error"
                    }
                }
            }
        },
        "/api/agrupacio": {
            "post": {
                "tags": [
                    "Agrupacions"
                ],
                "summary": "Inserir una nova agrupacio.",
                "operationId": "App\\Http\\Controllers\\AgrupacioController::crearAgrupacio",
                "parameters": [
                    {
                        "name": "nom",
                        "in": "query",
                        "description": "Nom de l'agrupacio",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "descripcio",
                        "in": "query",
                        "description": "Descripcio de l'agrupacio",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna l'agrupacio que hem inserit."
                    },
                    "default": {
                        "description": "S'ha produit un error."
                    }
                }
            }
        },
...
        "/api/bloquejats": {
            "get": {
                "tags": [
                    "Bloquejats"
                ],
                "summary": "Veure tots els usuaris bloquejats",
                "description": "",
                "operationId": "App\\Http\\Controllers\\BloquejatController::getBloquejats",
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "500": {
                        "description": "S'ha produit un error"
                    }
                }
            }
        },
        "/api/bloquejar": {
            "post": {
                "tags": [
                    "Bloquejats"
                ],
                "summary": "Inserir un nou bloquejat.",
                "operationId": "App\\Http\\Controllers\\BloquejatController::bloquejar",
                "parameters": [
                    {
                        "name": "bloquejat_id",
                        "in": "query",
                        "description": "id del bloquejat",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "bloquejador_id",
                        "in": "query",
                        "description": "id de la persona",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna l'usuari que hem bloquejat."
                    },
                    "default": {
                        "description": "S'ha produit un error."
                    }
                }
            }
        },
...
        "/api/pertanyenballarins": {
            "get": {
                "tags": [
                    "Ballarins"
                ],
                "summary": "Veure totes els ballarins.",
                "operationId": "App\\Http\\Controllers\\PertanyenBallariController::getPertanyen_ballarins",
                "responses": {
                    "200": {
                        "description": "Retorna tots els ballarins."
                    },
                    "default": {
                        "description": "S'ha produit un error."
                    }
                }
            }
        },
        "/api/pertanyenballari": {
            "post": {
                "tags": [
                    "Ballarins"
                ],
                "summary": "Inserir una nou ballari.",
                "operationId": "App\\Http\\Controllers\\PertanyenBallariController::pertanyen_ballari",
                "parameters": [
                    {
                        "name": "entitat_id",
                        "in": "query",
                        "description": "Id de l'entitat",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "ballari_id",
                        "in": "query",
                        "description": "Id del ballari",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna el ballari que hem inserit."
                    },
                    "default": {
                        "description": "S'ha produit un error."
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}